Move LayoutRPC to script_layout_interface.
