+++ components/layout/query.rs
@@ -16,9 +16,11 @@ use gfx::display_list::OpaqueNode;
 use gfx_traits::LayerId;
 use layout_thread::LayoutThreadData;
 use opaque_node::OpaqueNodeMethods;
-use script::layout_interface::{ContentBoxResponse, NodeOverflowResponse, ContentBoxesResponse, NodeGeometryResponse};
-use script::layout_interface::{HitTestResponse, LayoutRPC, OffsetParentResponse, NodeLayerIdResponse};
-use script::layout_interface::{ResolvedStyleResponse, MarginStyleResponse};
+use script_layout_interface::rpc::{ContentBoxResponse, ContentBoxesResponse};
+use script_layout_interface::rpc::{HitTestResponse, LayoutRPC};
+use script_layout_interface::rpc::{MarginStyleResponse, NodeGeometryResponse};
+use script_layout_interface::rpc::{NodeLayerIdResponse, NodeOverflowResponse};
+use script_layout_interface::rpc::{OffsetParentResponse, ResolvedStyleResponse};
 use script_layout_interface::wrapper_traits::{LayoutNode, ThreadSafeLayoutNode};
 use script_traits::LayoutMsg as ConstellationMsg;
 use script_traits::UntrustedNodeAddress;
