+++ components/script_layout_interface/Cargo.toml
@@ -9,8 +9,10 @@ name = "script_layout_interface"
 path = "lib.rs"
 
 [dependencies]
+app_units = {version = "0.2.3", features = ["plugins"]}
 bitflags = "0.7"
 canvas_traits = {path = "../canvas_traits"}
+euclid = {version = "0.6.4", features = ["plugins"]}
 gfx_traits = {path = "../gfx_traits"}
 heapsize = "0.3.0"
 heapsize_plugin = "0.1.2"
@@ -19,6 +21,7 @@ libc = "0.2"
 msg = {path = "../msg"}
 plugins = {path = "../plugins"}
 range = {path = "../range"}
+script_traits = {path = "../script_traits"}
 selectors = {version = "0.6", features = ["heap_size"]}
 string_cache = {version = "0.2.20", features = ["heap_size"]}
 style = {path = "../style"}
